feat(symthaea): cognitive loop refinements, mesh radio, voice, demo UI
Cognitive loop:
- Ethics engine, feedback state, cycle strategy, threshold, accessor updates
- Drive manager, feedback helpers, cycle extracted improvements
- Cycle phase dynamics/output refinements

Mind & Mesh:
- Mesh radio enhancements (bridge, receiver, mod)
- Mind mesh, mod, tick updates

Voice:
- Articulatory synthesizer, formant targets, orchestrator, vocoder updates

Infrastructure:
- Demo UI (static/demo.js, index.html, style.css)
- API demo_runner updates
- Nightly CI workflow update
- Papers: HAI paper TeX, psych-bench paper/supplement PDFs
- Integration test expansion

+++ symthaea/src/cognitive_loop/helpers/mod.rs
@@ -534,7 +534,7 @@ impl CognitiveLoopService {
         self.stats.adaptive_speech_rate = self.adaptive_behavior.speech_rate_multiplier;
 
         // Prediction confidence stats
-        self.stats.prediction_confidence = self.prediction_confidence;
+        self.stats.prediction_confidence = self.prediction_confidence as f32;
         // Decay rate: higher when in uncertain states
         self.stats.confidence_decay_rate = match self.adaptive_behavior.action_hint {
             ActionHint::Stabilize | ActionHint::SeekInput => 0.05,
